Add local development changes and some corrections
