+++ .env
@@ -1,0 +1,3 @@
+CONTAINER_NAME=go-tpcc
+VERSION=latest
+REPOSITORY=lockecodes
