+++ Dockerfile
@@ -1,4 +1,12 @@
-FROM ubuntu:18.04
-COPY go-tpcc /go-tpcc
-RUN chmod +rx /go-tpcc
-ENTRYPOINT ["/go-tpcc"]
+FROM bitnami/golang:1.15
+
+WORKDIR /opt/bitnami/go/src/go-tpcc
+COPY src .
+RUN go build .
+COPY entrypoint.sh .
+RUN chmod +x entrypoint.sh \
+    && mkdir -p /var/secrets/certs
+
+ENV CERTIFICATE_FILE_STR ''
+ENV DEBUG false
+ENTRYPOINT ["/opt/bitnami/go/src/go-tpcc/entrypoint.sh"]
