+++ docker-compose.dev.yaml
@@ -1,0 +1,17 @@
+version: '3.9'
+services:
+  tpcc:
+    entrypoint: bash
+    #    depends_on:
+#      - pg
+    volumes:
+      - ./src:/opt/bitnami/go/src/go-tpcc
+#  pg:
+#    image: postgres:12
+#    ports:
+#      - '5432:5432'
+#    environment:
+#      POSTGRES_PASSWORD: test
+#      POSTGRES_HOST_AUTH_METHOD: trust
+#      POSTGRES_DB: test
+#      POSTGRES_USER: test
