--- go.mod
@@ -1,13 +1,0 @@
-module github.com/Percona-Lab/go-tpcc
-
-require (
-	github.com/go-sql-driver/mysql v1.5.0
-	github.com/jackc/pgconn v1.7.0
-	github.com/jackc/pgx/v4 v4.9.0
-	github.com/mitchellh/go-homedir v1.1.0
-	github.com/spf13/cobra v1.0.0
-	github.com/spf13/viper v1.7.1
-	go.mongodb.org/mongo-driver v1.4.2
-)
-
-go 1.15
